+++ components/Blogs/ViewCard.tsx
