+++ package.json
@@ -11,9 +11,11 @@
   },
   "dependencies": {
     "bulma": "^0.9.4",
+    "easymde": "^2.17.0",
     "next": "^12.2.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-simplemde-editor": "^5.0.2",
     "sass": "^1.54.4"
   },
   "devDependencies": {
