+++ pages/blogs/create.tsx
@@ -1,19 +1,63 @@
+
 import type {NextPage} from 'next'
+import React, {useCallback, useState, useMemo} from 'react'
+import 'easymde/dist/easymde.min.css'
+import dynamic from 'next/dynamic'
 
-import CreateBlog from '@/components/Blogs/CreateBlogs'
+const BlogPage: NextPage = () => {
+    const SimpleMdeReact = dynamic(async () => import('react-simplemde-editor'), {ssr: false})
 
-const BlogPage: NextPage = () => (
-    <section className='hero has-background-white-bis is-fullheight-with-navbar'>
-        <div className='hero-body'>
+    const [value, setValue] = useState('Initial value')
 
-            <div className='container box'>
-                <div className='mb-6'>
-                    <p className='title is-1'>Add Blog</p>
+    const onChange = useCallback((value: string) => {
+        setValue(value)
+    }, [])
+
+    const autofocusNoSpellcheckerOptions = useMemo(() => ({
+        autofocus: true,
+        spellChecker: false,
+    } as SimpleMdeReact.Options), [])
+
+    return (
+        <section className='hero has-background-white-bis is-fullheight-with-navbar'>
+
+            <div className='hero-body'>
+                <div className='container'>
+                    <div className='columns'>
+                        <div className='column is-half is-offset-3'>
+                            <div className='card'>
+                                <div className='card-header'>
+                                    <p className='card-header-title'>Create Blog</p>
+                                </div>
+                                <div className='card-content'>
+                                    <form action=''>
+                                        <div className='field'>
+                                            <label htmlFor='title' className='label'>Title</label>
+                                            <div className='control'>
+                                                <input type='text' className='input' id='title' />
+                                            </div>
+                                        </div>
+                                        <div className='field'>
+                                            <label htmlFor='message' className='label'>Message</label>
+                                            {/* <div className='control'>
+                                                <textarea className='textarea' id='message' style={{minHeight: '20em'}}></textarea>
+                                            </div> */}
+                                            <SimpleMdeReact options={autofocusNoSpellcheckerOptions} value={value} onChange={onChange} />
+                                        </div>
+                                        <div className='field'>
+                                            <div className='control'>
+                                                <input type='submit' value='Submit' className='button is-primary is-outlined is-fullwidth' />
+                                            </div>
+                                        </div>
+                                    </form>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
                 </div>
-                <CreateBlog />
             </div>
-        </div>
-    </section>
-)
+        </section>
+    )
+}
 
 export default BlogPage
